chore(utils): add helper ignore method to clean up warnings
