+++ internal/utils/exec.go
@@ -2,8 +2,6 @@ package utils
 
 import (
 	"bytes"
-	"io"
-	"os"
 	"os/exec"
 
 	"github.com/pkg/errors"
@@ -24,25 +22,5 @@ func ExecCommand(executable string, args ...string) error {
 		return errors.Wrap(err, stdout.String())
 	}
 
-	return nil
-}
-
-// MoveFile physically moves file from source path to destination path
-// If dest already exists, MoveFile replaces it
-func MoveFile(source, dest string) error {
-	output, _ := os.Create(dest)
-	defer output.Close()
-
-	input, _ := os.Open(source)
-	_, err := io.Copy(output, input)
-	_ = input.Close()
-	if err != nil {
-		return errors.Wrapf(err, "failed to move from: %s to: %s", source, dest)
-	}
-
-	err = os.Remove(source)
-	if err != nil {
-		return err
-	}
 	return nil
 }
