+++ internal/utils/files.go
@@ -1,0 +1,28 @@
+package utils
+
+import (
+	"io"
+	"os"
+
+	"github.com/pkg/errors"
+)
+
+// MoveFile physically moves file from source path to destination path
+// If dest already exists, MoveFile replaces it
+func MoveFile(source, dest string) error {
+	output, _ := os.Create(dest)
+	defer Ignore(output.Close)
+
+	input, _ := os.Open(source)
+	_, err := io.Copy(output, input)
+	_ = input.Close()
+	if err != nil {
+		return errors.Wrapf(err, "failed to move from: %s to: %s", source, dest)
+	}
+
+	err = os.Remove(source)
+	if err != nil {
+		return err
+	}
+	return nil
+}
