+++ internal/utils/ignore.go
@@ -1,0 +1,6 @@
+package utils
+
+// Ignore invokes handler and explicitly ignores error
+func Ignore(handler func() error) {
+	_ = handler()
+}
